Finished SpriteBatcher
SpriteBatcher uses geometry shaders
Utilities is now core only
VertexArrays now contains the state of the vao binding
BMP decoding now supports 2 more headers (56 and 52 bytes)
OBJ Decoding now supports no object's inside of .obj
Added in test scene

+++ src/Core/Input/Mouse.cpp
@@ -16,7 +16,7 @@ void Mouse::attachWindow(GLFWwindow* window) {
 Vector2<int> Mouse::getMousePosition() {  
 	double x, y;
 	glfwGetCursorPos(window, &x, &y);
-	return Vector2<int>(x, y);
+	return Vector2<int>(static_cast<int>(x), static_cast<int>(y));
 }
 
 std::pair<KEY, KEY> Mouse::getMouseKey() {
